requirements.txt
speechrecognition
gtts
playsound
opencv-python
wikipedia
requests
selenium
webdriver-manager
python-docx
youtube-search
cython
numpy
